Move custom attr_readers to separated Extension
