+++ lib/luna_park.rb
@@ -22,6 +22,8 @@ require_relative 'luna_park/extensions/serializable'
 require_relative 'luna_park/extensions/comparable_debug'
 require_relative 'luna_park/extensions/comparable'
 require_relative 'luna_park/extensions/validatable'
+require_relative 'luna_park/extensions/predicate_attr_accessor'
+require_relative 'luna_park/extensions/coercible_attr_accessor'
 require_relative 'luna_park/extensions/dsl/attributes'
 require_relative 'luna_park/extensions/dsl/foreign_key'
 require_relative 'luna_park/entities/simple'
